Refactor xperiments sequence
Renamed spacetime data and engine modules for granule, scalar, and vector fields to new naming conventions (A/B/C). Updated all relevant imports in validation and experiment launcher scripts to match the new module names. Removed the obsolete _launcher_archive.py file from F_vector_field experiments.
