+++ openwave/validations/stability_analysis_springmass.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from openwave.common import constants, equations
 
-import openwave.spacetime.D_granule_based_data as data_grid
+import openwave.spacetime.A_granule_based_data as data_grid
 
 
 # From constants
